AG-12915 - Use docs page styles for error page
Remove ErrorPage react component and just use an Astro component
