--- documentation/ag-grid-docs/src/components/error/ErrorPage.module.scss
@@ -1,21 +1,0 @@
-@use 'design-system' as *;
-
-.errorPage {
-    height: 100%;
-    margin-top: $spacing-size-12;
-    margin-bottom: calc($spacing-size-12 * 4);
-
-    h2 {
-        margin-top: $spacing-size-8;
-    }
-}
-
-.content {
-    p {
-        margin-bottom: $spacing-size-2;
-    }
-
-    pre {
-        background-color: rgb(254 226 226); // TODO: Find better color
-    }
-}
