+++ documentation/ag-grid-docs/src/content/errors/2.mdoc
@@ -1,6 +1,12 @@
 ---
-title: Row data error
+title: "Row data error #2"
 description: Row data description
 ---
 
 This is an error about duplicate node ids. See [Row IDs docs](./row-ids/#row-ids).
+
+## Why is this error happening?
+
+Because you have a duplicate node id
+
+
