+++ documentation/ag-grid-docs/src/pages/[framework]-data-grid/errors/[code].astro
@@ -1,10 +1,17 @@
 ---
 import { ERRORS, type ErrorCode } from 'ag-grid-community';
 import Layout from '@layouts/Layout.astro';
-import { ErrorPage } from '@components/error/ErrorPage';
-import { getCollection, type CollectionEntry } from 'astro:content';
-import type { Framework } from '@ag-grid-types';
+import { getCollection, getEntry, type CollectionEntry } from 'astro:content';
+import type { Framework, MenuItem, MenuSection } from '@ag-grid-types';
 import { FRAMEWORKS } from '@constants';
+import { getFilteredMenuSections } from '@components/pages-navigation/utils/getFilteredMenuSections';
+import styles from '@ag-website-shared/components/page-styles/docs.module.scss';
+import { PagesNavigation } from '@components/pages-navigation/components/PagesNavigation';
+import { Header } from '@features/docs/components/Header';
+import { getErrorText } from '@utils/getErrorText';
+import { SideNavigation } from '@components/pages-navigation/components/SideNavigation';
+import { getHeadings } from '@utils/markdoc/getHeadings';
+import { DOCS_TAB_ITEM_ID_PREFIX } from '@ag-website-shared/constants';
 
 interface Params {
     framework: Framework;
@@ -35,14 +42,80 @@ export async function getStaticPaths() {
 
 const { framework, code } = Astro.params as Params;
 const { errorPage } = Astro.props as Props;
-const title = errorPage?.data.title ? errorPage?.data.title : `AG Grid error #${code}`;
+const title = errorPage?.data.title ? errorPage?.data.title : `AG Grid Error #${code}`;
 const description = errorPage?.data.description ? errorPage?.data.description : '';
 
 const { Content } = errorPage ? await errorPage.render() : { Content: undefined };
+
+const { data: menu } = await getEntry('menu', 'data');
+const menuSections: MenuSection[] = getFilteredMenuSections({
+    menuSections: menu.main.sections,
+    framework,
+});
+const path = Astro.url.pathname;
+const pageName = `errors/${code}`;
+const defaultErrorText = getErrorText({ errorCode: code });
+
+const getTabItemSlug = (id: string) => `${DOCS_TAB_ITEM_ID_PREFIX}${id}`;
+const fullErrorTextHeading = { slug: getTabItemSlug('full-error-text'), depth: 2, text: 'Full Error Text' };
+const headings = errorPage
+    ? (
+          await getHeadings({
+              title,
+              pageName,
+              markdocContent: errorPage.body,
+              framework,
+              getTabItemSlug,
+          })
+      ).concat(fullErrorTextHeading)
+    : [fullErrorTextHeading];
 ---
 
-<Layout title={title} description={description} showSearchBar={true} showDocsNav={false}>
-    <ErrorPage client:only="react" errorCode={code}>
-        {Content ? <Content framework={framework} /> : undefined}
-    </ErrorPage>
+<Layout title={title} description={description} showSearchBar={true} showDocsNav={true}>
+    <div class:list={[styles.contentViewport, 'layout-grid']}>
+        <PagesNavigation client:load menuSections={menuSections} framework={framework} pageName={pageName} />
+        <div
+            id="doc-content"
+            class:list={[
+                styles.docPage,
+                styles.errorPage,
+                {
+                    noSideMenu: styles.hideSideMenu,
+                },
+            ]}
+        >
+            <Header
+                client:load
+                title={title}
+                framework={framework}
+                path={path}
+                menuItems={menuSections as MenuItem[]}
+            />
+            <div class={styles.pageSections}>
+                {Content ? <Content framework={framework} /> : undefined}
+
+                <h2 id={fullErrorTextHeading.slug}>{fullErrorTextHeading.text}</h2>
+                <pre id="errorCodeText">{defaultErrorText}</pre>
+            </div>
+        </div>
+
+        <SideNavigation client:load headings={headings} />
+    </div>
 </Layout>
+
+<script>
+    import { getErrorText } from '@utils/getErrorText';
+
+    // Get errorCode from url, so it doesn't need to be passed
+    // in from Astro and we can import in this script
+    const errorCode = window.location.pathname.split('/').slice(-1)[0];
+    const searchParams = new URLSearchParams(window.location.search);
+    const params = Object.fromEntries(searchParams.entries());
+    const errorText = getErrorText({ errorCode, params });
+
+    const errorCodeTextEl = document.getElementById('errorCodeText');
+
+    if (errorCodeTextEl) {
+        errorCodeTextEl.textContent = errorText;
+    }
+</script>
