+++ external/ag-website-shared/src/components/page-styles/docs.module.scss
@@ -201,3 +201,9 @@ blockquote:where(:not(:last-child)) {
 .sideMenu {
     padding-top: $spacing-size-8 + $spacing-size-1;
 }
+
+.errorPage {
+    pre {
+        background-color: rgb(254 226 226); // TODO: Find better color
+    }
+}
